Move testing utils to package
This will be useful for testing in dependant packages
