+++ tests/__init__.py
@@ -1,0 +1,3 @@
+import dotenv
+
+dotenv.load_dotenv(dotenv.find_dotenv('.test.env'))  # noqa
